pyproject.toml
[tool.poetry]
name = "routeros-api"
version = "0.1.0"
description = ""
authors = ["Ferdinand Niedermann <nerdinand@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
laiarturs-ros-api = "^1.0.0"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
